requirements/base.txt
appdirs==1.4.3
djangorestframework==3.8.2
Markdown==2.6.11
django-rest-swagger==2.1.1
django-cors-headers==2.3.0
django-filter==1.1.0
Django==2.0.7
django-ses==0.8.5
pika==0.12.0
python-docx==0.8.6
djangorestframework-filters==0.10.2
djangorestframework-bulk==0.2.1
celery==4.2.0
suds-py3==1.3.3.0
xmltodict==0.11.0
djangorestframework-xml==1.3.0
boto3==1.7.56
inflection==0.3.1
python-stdnum==1.9
psycopg2-binary==2.7.5
Pillow==5.2.0
